src/test/java/io/novaordis/utilities/os/OSTypeTest.java
/*
 * Copyright (c) 2017 Nova Ordis LLC
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *    http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */

package io.novaordis.utilities.os;

import org.junit.Test;

import static org.junit.Assert.assertEquals;
import static org.junit.Assert.fail;

/**
 * @author Ovidiu Feodorov <ovidiu@novaordis.com>
 * @since 6/6/17
 */
public class OSTypeTest {

    // Constants -------------------------------------------------------------------------------------------------------

    // Static ----------------------------------------------------------------------------------------------------------

    // Attributes ------------------------------------------------------------------------------------------------------

    // Constructors ----------------------------------------------------------------------------------------------------

    // Public ----------------------------------------------------------------------------------------------------------

    // Tests -----------------------------------------------------------------------------------------------------------

    @Test
    public void knownValues() {

        OSType t = OSType.getCurrent();

        if (OSType.MAC.equals(t) || OSType.LINUX.equals(t) || OSType.WINDOWS.equals(t)) {

            return;
        }

        fail("unknown OSType " + t);
    }

    @Test
    public void reset() {

        OSType crt = OSType.getCurrent();

        try {

            OSType.current = OSType.TEST;

            assertEquals(OSType.TEST, OSType.getCurrent());

            OSType.reset();

            assertEquals(crt, OSType.getCurrent());

        }
        finally {

            //
            // restore anyway
            //

            OSType.reset();
        }
    }

    // Package protected -----------------------------------------------------------------------------------------------

    // Protected -------------------------------------------------------------------------------------------------------

    // Private ---------------------------------------------------------------------------------------------------------

    // Inner classes ---------------------------------------------------------------------------------------------------

}
